moving back  some widgets to main.dart
